refactor: extract PR functions to github/pr.py (#139)
